+++ joclJetBackgroundCutModule/src/jetBackgroundCut.h
@@ -2,72 +2,70 @@
 #define JETBACKGROUNDCUT_H
 
 #include <fun4all/SubsysReco.h>
-#include <string>
-#include <cmath>
-#include <phool/recoConsts.h>
 #include <globalvertex/GlobalVertex.h>
 #include <phparameter/PHParameters.h>
+#include <cmath>
+#include <string>
 class PHCompositeNode;
 class CentralityInfo;
 class jetBackgroundCut : public SubsysReco
 {
  public:
+  explicit jetBackgroundCut(const std::string &jetNodeName, const std::string &name = "jetBackgroundCutModule", int debug = 0, bool doAbort = false, GlobalVertex::VTXTYPE vtxtype = GlobalVertex::MBD, int sysvar = 0);
 
-  jetBackgroundCut(const std::string jetNodeName, const std::string &name = "jetBackgroundCutModule", const int debug = 0, const bool doAbort = 0, GlobalVertex::VTXTYPE vtxtype = GlobalVertex::MBD, int sysvar = 0);
+  ~jetBackgroundCut() override;
 
-  virtual ~jetBackgroundCut();
-
-  bool failsLoEmFracETCut(float emFrac, float ET, bool dPhiCut, bool isDijet)
+  bool failsLoEmFracETCut(float emFrac, float ET, bool dPhiCut, bool isDijet) const
   {
     float widen = 0;
     float shift = 0;
-    if(_sysvar > 0)
-      {
-	widen = 0.05;
-	shift = -2.5;
-      }
-    else if(_sysvar < 0)
-      {
-	widen = -0.05;
-	shift = 2.5;
-      }
-    return (emFrac < (0.1+widen) && ET > (50*emFrac+20+shift)) && (dPhiCut || !isDijet);
+    if (_sysvar > 0)
+    {
+      widen = 0.05;
+      shift = -2.5;
+    }
+    else if (_sysvar < 0)
+    {
+      widen = -0.05;
+      shift = 2.5;
+    }
+    return (emFrac < (0.1 + widen) && ET > (50 * emFrac + 20 + shift)) && (dPhiCut || !isDijet);
   }
 
-  bool failsHiEmFracETCut(float emFrac, float ET, bool dPhiCut, bool isDijet)
+  bool failsHiEmFracETCut(float emFrac, float ET, bool dPhiCut, bool isDijet) const
   {
     float widen = 0;
     float shift = 0;
-    if(_sysvar > 0)
-      {
-	widen = -0.05;
-	shift = -2.5;
-      }
-    else if(_sysvar < 0)
-      {
-	widen = 0.05;
-	shift = 2.5;
-      }
-    return (emFrac > (0.9+widen) && ET > (-50*emFrac+70+shift)) && (dPhiCut || !isDijet);
+    if (_sysvar > 0)
+    {
+      widen = -0.05;
+      shift = -2.5;
+    }
+    else if (_sysvar < 0)
+    {
+      widen = 0.05;
+      shift = 2.5;
+    }
+    return (emFrac > (0.9 + widen) && ET > (-50 * emFrac + 70 + shift)) && (dPhiCut || !isDijet);
   }
 
-  bool failsIhFracCut(float emFrac, float ohFrac)
+  bool failsIhFracCut(float emFrac, float ohFrac) const
   {
     float shift = 0;
-    if(_sysvar > 0)
-      {
-	shift = 0.05;
-      }
-    else if(_sysvar < 0)
-      {
-	shift = -0.05;
-      }
-    return emFrac + ohFrac < 0.65+shift;
+    if (_sysvar > 0)
+    {
+      shift = 0.05;
+    }
+    else if (_sysvar < 0)
+    {
+      shift = -0.05;
+    }
+    return emFrac + ohFrac < 0.65 + shift;
   }
 
   bool failsdPhiCut(float dPhi, bool isDijet)
   {
-    return dPhi < 3*M_PI/4 && isDijet;
+    return dPhi < 3 * M_PI / 4 && isDijet;
   }
 
   int Init(PHCompositeNode *topNode) override;
@@ -82,14 +80,14 @@ class jetBackgroundCut : public SubsysReco
 
   void Print(const std::string &what = "ALL") const override;
 
-  void CreateNodeTree(PHCompositeNode* topNode);
+  void CreateNodeTree(PHCompositeNode *topNode);
 
   void SetDefaultParams()
   {
-    _cutParams.set_int_param("failsLoEmJetCut",0);
-    _cutParams.set_int_param("failsHiEmJetCut",0);
-    _cutParams.set_int_param("failsIhJetCut",0);
-    _cutParams.set_int_param("failsAnyJetCut",0);
+    _cutParams.set_int_param("failsLoEmJetCut", 0);
+    _cutParams.set_int_param("failsHiEmJetCut", 0);
+    _cutParams.set_int_param("failsIhJetCut", 0);
+    _cutParams.set_int_param("failsAnyJetCut", 0);
   }
 
  private:
